# Optimise StackTraceSnippet.Get