--- a/benchmarks/MiniProfiler.Benchmarks/Program.cs
+++ b/benchmarks/MiniProfiler.Benchmarks/Program.cs
@@ -7,11 +7,6 @@ namespace Benchmarks
     {
         public static void Main(string[] args)
         {
-            //var creation = new Benchmarks.CreationBenchmarks();
-            //while (true)
-            //{
-            //    creation.StartStopProfiler();
-            //}
             BenchmarkSwitcher.FromAssembly(typeof(Program).GetTypeInfo().Assembly).Run(args);
         }
     }